requirements.txt
fastapi
ipykernel
joblib
matplotlib
mlflow
numpy
optuna
pandas
psycopg2-binary
pytorch-forecasting
pytorch-lightning
SQLAlchemy
torch
torchmetrics
torchvision
tqdm
uvicorn